KernelSU: Track KSU-Next versioning for universal sucompat toggle visibility
KernelSU-Next Manager arbitrarily chooses to guard the sucompat toggle
behind a minimum version requirement.

As a result, I'm forced to sync with KSU-Next versioning in order to
maintain feature parity and cross-compatibility between it and the
official KSU Manager, even though the sucompat toggle would function
perfectly well on both apps if not for this frivolous check.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>